♥ 바이너리로 구동보드 데이터 받게끔 수정, 기타 amcl 및 각속도 최적화

+++ tetraDS/src/drive_module.h
@@ -24,4 +24,4 @@ int drvm_read_parameter(int fd, int param, int *value);
 int drvm_read_odometry(int fd, double *Xpos_mm, double *Ypos_mm, double *deg);
 int drvm_set_Charge_On(int fd);
 int drvm_set_Charge_Off(int fd);
-int drvm_set_velocity2(int fd, int left, int right, int *x, int *y, int *theta, int *bumper, int *emg);
+int drvm_set_velocity2(int fd, int left, int right, double *x, double *y, double *theta, int *bumper, int *emg);
