+++ tetraDS/src/dssp_rs232_drive_module.c
@@ -169,7 +169,7 @@ int  dssp_rs232_drv_module_set_charge(int mode)
 	return ret;
 }
 
-int  dssp_rs232_drv_module_set_velocity2(int velocity_l, int velocity_r, int *Xpos_mm, int *Ypos_mm, int *deg, int *bumper, int *emg)
+int  dssp_rs232_drv_module_set_velocity2(int velocity_l, int velocity_r, double *Xpos_mm, double *Ypos_mm, double *deg, int *bumper, int *emg)
 {
 	int ret;
